Sun, Sep 1, 2024, 5:40 PM -04:00

+++ python /week1/score.py
@@ -1,4 +1,4 @@
-score = int(input("score: "))
+score = float(input("score: "))
 if 90 <= score <=100:
     print("Grade A")
 elif score >=80:
